+++ src/ECS/Components/MouvementComponent.cpp
@@ -11,12 +11,12 @@
 
 /**
  * It's a constructor that takes three parameters and assigns them to the three member variables
- * 
+ *
  * @param dirX The direction of the movement on the X axis.
  * @param dirY The direction of the movement on the Y axis.
  * @param speed the speed of the entity
  */
-MouvementComponent::MouvementComponent(float dirX, float dirY, float speed)
+MouvementComponent::MouvementComponent(int dirX, int dirY, float speed)
     : _dirX(dirX)
     , _dirY(dirY)
     , _speed(speed)
@@ -30,7 +30,7 @@ MouvementComponent::~MouvementComponent() noexcept {}
 
 /**
  * It returns the value of the private member _dirX
- * 
+ *
  * @return The direction of the x axis.
  */
 float MouvementComponent::getDirX() const noexcept
@@ -40,7 +40,7 @@ float MouvementComponent::getDirX() const noexcept
 
 /**
  * It returns the value of the private attribute _dirY
- * 
+ *
  * @return The direction of the Y axis.
  */
 float MouvementComponent::getDirY() const noexcept
@@ -50,7 +50,7 @@ float MouvementComponent::getDirY() const noexcept
 
 /**
  * It returns the speed of the entity
- * 
+ *
  * @return The speed of the object.
  */
 float MouvementComponent::getSpeed() const noexcept
@@ -60,27 +60,27 @@ float MouvementComponent::getSpeed() const noexcept
 
 /**
  * It sets the direction of the entity on the X axis
- * 
- * @param x The x-coordinate of the direction vector.
+ *
+ * @param x The x-coordinate of the direction vector. (-1, 1 or 0)
  */
-void MouvementComponent::setDirX(float x) noexcept
+void MouvementComponent::setDirX(int x) noexcept
 {
     _dirX = x;
 }
 
 /**
  * It sets the direction of the Y axis
- * 
- * @param y The y-axis direction of the movement.
+ *
+ * @param y The y-axis direction of the movement. (-1, 1 or 0)
  */
-void MouvementComponent::setDirY(float y) noexcept
+void MouvementComponent::setDirY(int y) noexcept
 {
     _dirY = y;
 }
 
 /**
  * It sets the speed of the entity
- * 
+ *
  * @param speed The speed of the entity.
  */
 void MouvementComponent::setSpeed(float speed) noexcept
